@@ -35,7 +35,7 @@
         "colinmollenhour/php-redis-session-abstract": "~1.4.0",
         "composer/composer": "^1.6",
         "elasticsearch/elasticsearch": "~2.0||~5.1||~6.1",
-        "magento/composer": "~1.5.0",
+        "magento/composer": "1.6.x-dev",
         "magento/magento-composer-installer": ">=0.1.11",
         "magento/zendframework1": "~1.14.2",
         "monolog/monolog": "^1.17",
@@ -46,9 +46,9 @@
         "phpseclib/mcrypt_compat": "1.0.8",
         "phpseclib/phpseclib": "2.0.*",
         "ramsey/uuid": "~3.8.0",
-        "symfony/console": "~4.1.0||~4.2.0||~4.3.0",
-        "symfony/event-dispatcher": "~4.1.0||~4.2.0||~4.3.0",
-        "symfony/process": "~4.1.0||~4.2.0||~4.3.0",
+        "symfony/console": "~4.4.0",
+        "symfony/event-dispatcher": "~4.4.0",
+        "symfony/process": "~4.4.0",
         "tedivm/jshrink": "~1.3.0",
         "tubalmartin/cssmin": "4.1.1",
         "webonyx/graphql-php": "^0.13.8",
