@@ -27,8 +27,8 @@
         "magento/zendframework1": "~1.14.2",
         "monolog/monolog": "^1.17",
         "wikimedia/less.php": "~1.8.0",
-        "symfony/console": "~4.1.0",
-        "symfony/process": "~4.1.0",
+        "symfony/console": "~4.4.0",
+        "symfony/process": "~4.4.0",
         "tedivm/jshrink": "~1.3.0",
         "zendframework/zend-code": "~3.3.0",
         "zendframework/zend-crypt": "^2.6.0",
