@@ -12,6 +12,8 @@ use Magento\Framework\Data\Form\FormKey\Validator;
 use Magento\Framework\View\Result\PageFactory;
 
 /**
+ * View products compare in frontend
+ *
  * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  */
 class Index extends \Magento\Catalog\Controller\Product\Compare implements HttpGetActionInterface
@@ -74,22 +76,11 @@ class Index extends \Magento\Catalog\Controller\Product\Compare implements HttpG
      */
     public function execute()
     {
-        $items = $this->getRequest()->getParam('items');
-
         $beforeUrl = $this->getRequest()->getParam(self::PARAM_NAME_URL_ENCODED);
         if ($beforeUrl) {
             $this->_catalogSession->setBeforeCompareUrl(
                 $this->urlDecoder->decode($beforeUrl)
             );
-        }
-
-        if ($items) {
-            $items = explode(',', $items);
-            /** @var \Magento\Catalog\Model\Product\Compare\ListCompare $list */
-            $list = $this->_catalogProductCompareList;
-            $list->addProducts($items);
-            $resultRedirect = $this->resultRedirectFactory->create();
-            return $resultRedirect->setPath('*/*/*');
         }
         return $this->resultPageFactory->create();
     }
