@@ -14,6 +14,7 @@ use Magento\Catalog\Model\ResourceModel\Product\Compare\Item\Collection;
  * @api
  * @SuppressWarnings(PHPMD.LongVariable)
  * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
+ * @SuppressWarnings(PHPMD.CookieAndSessionMisuse)
  * @since 100.0.2
  */
 class Compare extends \Magento\Framework\Url\Helper\Data
@@ -145,16 +146,9 @@ class Compare extends \Magento\Framework\Url\Helper\Data
      */
     public function getListUrl()
     {
-        $itemIds = [];
-        foreach ($this->getItemCollection() as $item) {
-            $itemIds[] = $item->getId();
-        }
-
         $params = [
-            'items' => implode(',', $itemIds),
             \Magento\Framework\App\ActionInterface::PARAM_NAME_URL_ENCODED => $this->getEncodedUrl()
         ];
-
         return $this->_getUrl('catalog/product_compare', $params);
     }
 
