@@ -7,7 +7,7 @@
  * @api
  */
 define([
-    '../model/quote'
+    'Magento_Checkout/js/model/quote'
 ], function (quote) {
     'use strict';
 
