@@ -10,7 +10,7 @@ namespace Magento\Downloadable\Block\Adminhtml\Catalog\Product\Edit\Tab\Download
  *
  * @author      Magento Core Team <core@magentocommerce.com>
  *
- * @deprecated
+ * @deprecated because of new class which adds grids samples
  * @see \Magento\Downloadable\Ui\DataProvider\Product\Form\Modifier\Samples
  */
 class Samples extends \Magento\Backend\Block\Widget
@@ -250,7 +250,7 @@ class Samples extends \Magento\Backend\Block\Widget
      */
     public function getConfigJson()
     {
-        $url = $this->_urlFactory->create()->addSessionParam()->getUrl(
+        $url = $this->_urlFactory->create()->getUrl(
             'adminhtml/downloadable_file/upload',
             ['type' => 'samples', '_secure' => true]
         );
