@@ -16,7 +16,7 @@ use Magento\Framework\UrlInterface;
 use Magento\Framework\Stdlib\ArrayManager;
 
 /**
- * Class SamplesTest
+ * Test for class Magento\Downloadable\Ui\DataProvider\Product\Form\Modifier\Samples
  * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  */
 class SamplesTest extends \PHPUnit\Framework\TestCase
@@ -141,7 +141,7 @@ class SamplesTest extends \PHPUnit\Framework\TestCase
             ->method('isSingleStoreMode');
         $this->typeUploadMock->expects($this->once())
             ->method('toOptionArray');
-        $this->urlBuilderMock->expects($this->once())
+        $this->urlBuilderMock->expects($this->never())
             ->method('addSessionParam')
             ->willReturnSelf();
         $this->urlBuilderMock->expects($this->once())
