@@ -69,9 +69,6 @@ class AbstractTemplateTest extends \PHPUnit\Framework\TestCase
         $this->design = $this->getMockBuilder(\Magento\Framework\View\DesignInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $this->registry = $this->getMockBuilder(\Magento\Framework\Registry::class)
-            ->disableOriginalConstructor()
-            ->getMock();
         $this->appEmulation = $this->getMockBuilder(\Magento\Store\Model\App\Emulation::class)
             ->disableOriginalConstructor()
             ->getMock();
@@ -177,7 +174,7 @@ class AbstractTemplateTest extends \PHPUnit\Framework\TestCase
             ->disableOriginalConstructor()
             ->getMock();
 
-        $filterTemplate->expects($this->once())
+        $filterTemplate->expects($this->never())
             ->method('setUseSessionInUrl')
             ->with(false)
             ->will($this->returnSelf());
@@ -253,7 +250,6 @@ class AbstractTemplateTest extends \PHPUnit\Framework\TestCase
         $filterTemplate = $this->getMockBuilder(\Magento\Email\Model\Template\Filter::class)
             ->setMethods(
                 [
-                'setUseSessionInUrl',
                 'setPlainTemplateMode',
                 'setIsChildTemplate',
                 'setDesignParams',
@@ -267,9 +263,7 @@ class AbstractTemplateTest extends \PHPUnit\Framework\TestCase
             )
             ->disableOriginalConstructor()
             ->getMock();
-        $filterTemplate->expects($this->once())
-            ->method('setUseSessionInUrl')
-            ->will($this->returnSelf());
+
         $filterTemplate->expects($this->once())
             ->method('setPlainTemplateMode')
             ->will($this->returnSelf());
