@@ -9,18 +9,19 @@ define([
     'uiComponent',
     'Magento_Checkout/js/model/quote',
     'Magento_SalesRule/js/action/set-coupon-code',
-    'Magento_SalesRule/js/action/cancel-coupon'
-], function ($, ko, Component, quote, setCouponCodeAction, cancelCouponAction) {
+    'Magento_SalesRule/js/action/cancel-coupon',
+    'Magento_SalesRule/js/model/coupon'
+], function ($, ko, Component, quote, setCouponCodeAction, cancelCouponAction, coupon) {
     'use strict';
 
     var totals = quote.getTotals(),
-        couponCode = ko.observable(null),
-        isApplied;
+        couponCode = coupon.getCouponCode(),
+        isApplied = coupon.getIsApplied();
 
     if (totals()) {
         couponCode(totals()['coupon_code']);
     }
-    isApplied = ko.observable(couponCode() != null);
+    isApplied(couponCode() != null);
 
     return Component.extend({
         defaults: {
