@@ -11,11 +11,7 @@ use Magento\Framework\App\Action\Action;
 use Magento\Framework\App\Action\Context;
 use Magento\Framework\App\Action\HttpGetActionInterface;
 use Magento\Framework\App\Action\HttpPostActionInterface;
-use Magento\Framework\App\ResponseInterface;
-use Magento\Framework\Controller\ResultInterface;
 use Magento\Framework\Exception\NoSuchEntityException;
-use Magento\Framework\Session\Generic as Session;
-use Magento\Framework\Session\SidResolverInterface;
 use Magento\Store\Api\StoreRepositoryInterface;
 use Magento\Store\Api\StoreResolverInterface;
 use Magento\Store\Model\Store;
@@ -23,7 +19,7 @@ use Magento\Store\Model\StoreResolver;
 use Magento\Store\Model\StoreSwitcher\HashGenerator;
 
 /**
- * Builds correct url to target store and performs redirect.
+ * Builds correct url to target store (group) and performs redirect.
  */
 class Redirect extends Action implements HttpGetActionInterface, HttpPostActionInterface
 {
@@ -38,16 +34,6 @@ class Redirect extends Action implements HttpGetActionInterface, HttpPostActionI
     private $storeResolver;
 
     /**
-     * @var SidResolverInterface
-     */
-    private $sidResolver;
-
-    /**
-     * @var Session
-     */
-    private $session;
-
-    /**
      * @var HashGenerator
      */
     private $hashGenerator;
@@ -56,30 +42,28 @@ class Redirect extends Action implements HttpGetActionInterface, HttpPostActionI
      * @param Context $context
      * @param StoreRepositoryInterface $storeRepository
      * @param StoreResolverInterface $storeResolver
-     * @param Session $session
-     * @param SidResolverInterface $sidResolver
+     * @param \Magento\Framework\Session\Generic $session
+     * @param \Magento\Framework\Session\SidResolverInterface $sidResolver
      * @param HashGenerator $hashGenerator
+     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
     public function __construct(
         Context $context,
         StoreRepositoryInterface $storeRepository,
         StoreResolverInterface $storeResolver,
-        Session $session,
-        SidResolverInterface $sidResolver,
+        \Magento\Framework\Session\Generic $session,
+        \Magento\Framework\Session\SidResolverInterface $sidResolver,
         HashGenerator $hashGenerator
     ) {
         parent::__construct($context);
         $this->storeRepository = $storeRepository;
         $this->storeResolver = $storeResolver;
-        $this->session = $session;
-        $this->sidResolver = $sidResolver;
         $this->hashGenerator = $hashGenerator;
     }
 
     /**
-     * Performs store redirect
+     * @inheritDoc
      *
-     * @return ResponseInterface|ResultInterface
      * @throws NoSuchEntityException
      */
     public function execute()
@@ -112,12 +96,6 @@ class Redirect extends Action implements HttpGetActionInterface, HttpPostActionI
                 StoreResolverInterface::PARAM_NAME => $targetStoreCode,
                 \Magento\Framework\App\ActionInterface::PARAM_NAME_URL_ENCODED => $encodedUrl,
             ];
-
-            if ($this->sidResolver->getUseSessionInUrl()) {
-                // allow customers to stay logged in during store switching
-                $sidName = $this->sidResolver->getSessionIdQueryParam($this->session);
-                $query[$sidName] = $this->session->getSessionId();
-            }
 
             $customerHash = $this->hashGenerator->generateHash($fromStore);
             $query = array_merge($query, $customerHash);
