@@ -278,6 +278,7 @@ class Store extends AbstractExtensibleModel implements
 
     /**
      * @var \Magento\Framework\Session\SidResolverInterface
+     * @deprecated Not used anymore.
      */
     protected $_sidResolver;
 
@@ -1199,7 +1200,6 @@ class Store extends AbstractExtensibleModel implements
      */
     public function getCurrentUrl($fromStore = true)
     {
-        $sidQueryParam = $this->_sidResolver->getSessionIdQueryParam($this->_getSession());
         $requestString = $this->_url->escape(ltrim($this->_request->getRequestString(), '/'));
 
         $storeUrl = $this->getUrl('', ['_secure' => $this->_storeManager->getStore()->isCurrentlySecure()]);
@@ -1218,12 +1218,6 @@ class Store extends AbstractExtensibleModel implements
         }
 
         $currQuery = $this->_request->getQueryValue();
-        if (isset($currQuery[$sidQueryParam])
-            && !empty($currQuery[$sidQueryParam])
-            && $this->_getSession()->getSessionIdForHost($storeUrl) != $currQuery[$sidQueryParam]
-        ) {
-            unset($currQuery[$sidQueryParam]);
-        }
 
         foreach ($currQuery as $key => $value) {
             $storeParsedQuery[$key] = $value;
