@@ -19,6 +19,9 @@ use Magento\Catalog\Model\Category;
 use Magento\Catalog\Model\Product;
 use Magento\Catalog\Model\ResourceModel\Product as ProductResource;
 use Magento\CatalogImportExport\Model\Import\Product\RowValidatorInterface;
+use Magento\CatalogInventory\Model\Stock;
+use Magento\CatalogInventory\Model\StockRegistry;
+use Magento\CatalogInventory\Model\StockRegistryStorage;
 use Magento\Framework\App\Bootstrap;
 use Magento\Framework\App\Filesystem\DirectoryList;
 use Magento\Framework\App\ObjectManager;
@@ -230,9 +233,9 @@ class ProductTest extends \Magento\TestFramework\Indexer\TestCase
         $existingProductIds = [$id1, $id2, $id3];
         $stockItems = [];
         foreach ($existingProductIds as $productId) {
-            /** @var $stockRegistry \Magento\CatalogInventory\Model\StockRegistry */
+            /** @var $stockRegistry StockRegistry */
             $stockRegistry = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->create(
-                \Magento\CatalogInventory\Model\StockRegistry::class
+                StockRegistry::class
             );
 
             $stockItem = $stockRegistry->getStockItem($productId, 1);
@@ -261,9 +264,9 @@ class ProductTest extends \Magento\TestFramework\Indexer\TestCase
 
         /** @var $stockItmBeforeImport \Magento\CatalogInventory\Model\Stock\Item */
         foreach ($stockItems as $productId => $stockItmBeforeImport) {
-            /** @var $stockRegistry \Magento\CatalogInventory\Model\StockRegistry */
+            /** @var $stockRegistry StockRegistry */
             $stockRegistry = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->create(
-                \Magento\CatalogInventory\Model\StockRegistry::class
+                StockRegistry::class
             );
 
             $stockItemAfterImport = $stockRegistry->getStockItem($productId, 1);
@@ -2031,9 +2034,9 @@ class ProductTest extends \Magento\TestFramework\Indexer\TestCase
         $this->_model->importData();
 
         foreach ($products as $sku => $manageStockUseConfig) {
-            /** @var \Magento\CatalogInventory\Model\StockRegistry $stockRegistry */
+            /** @var StockRegistry $stockRegistry */
             $stockRegistry = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->create(
-                \Magento\CatalogInventory\Model\StockRegistry::class
+                StockRegistry::class
             );
             $stockItem = $stockRegistry->getStockItemBySku($sku);
             $this->assertEquals($manageStockUseConfig, $stockItem->getUseConfigManageStock());
@@ -2940,5 +2943,29 @@ class ProductTest extends \Magento\TestFramework\Indexer\TestCase
             }
         }
         $this->assertEquals($expected, $actual);
+    }
+
+    /**
+     * Test that product stock status is updated after import
+     *
+     * @magentoDataFixture mediaImportImageFixture
+     * @magentoDataFixture Magento/Catalog/_files/product_simple.php
+     */
+    public function testProductStockStatusShouldBeUpdated()
+    {
+        /** @var $stockRegistry StockRegistry */
+        $stockRegistry = $this->objectManager->create(StockRegistry::class);
+        /** @var StockRegistryStorage $stockRegistryStorage */
+        $stockRegistryStorage = $this->objectManager->get(StockRegistryStorage::class);
+        $status = $stockRegistry->getStockStatusBySku('simple');
+        $this->assertEquals(Stock::STOCK_IN_STOCK, $status->getStockStatus());
+        $this->importDataForMediaTest('disable_product.csv');
+        $stockRegistryStorage->clean();
+        $status = $stockRegistry->getStockStatusBySku('simple');
+        $this->assertEquals(Stock::STOCK_OUT_OF_STOCK, $status->getStockStatus());
+        $this->importDataForMediaTest('enable_product.csv');
+        $stockRegistryStorage->clean();
+        $status = $stockRegistry->getStockStatusBySku('simple');
+        $this->assertEquals(Stock::STOCK_IN_STOCK, $status->getStockStatus());
     }
 }
