@@ -6,7 +6,11 @@
 namespace Magento\DownloadableImportExport\Model;
 
 use Magento\CatalogImportExport\Model\AbstractProductExportImportTestCase;
+use Magento\Catalog\Model\Product;
 
+/**
+ * Test export and import downloadable products
+ */
 class DownloadableTest extends AbstractProductExportImportTestCase
 {
     /**
@@ -17,15 +21,7 @@ class DownloadableTest extends AbstractProductExportImportTestCase
         return [
             'downloadable-product' => [
                 [
-                    'Magento/Downloadable/_files/product_downloadable.php'
-                ],
-                [
-                    'downloadable-product',
-                ],
-            ],
-            'downloadable-product-with-files' => [
-                [
-                    'Magento/Downloadable/_files/product_downloadable_with_files.php'
+                    'Magento/Downloadable/_files/product_downloadable_with_link_url_and_sample_url.php'
                 ],
                 [
                     'downloadable-product',
@@ -46,43 +42,54 @@ class DownloadableTest extends AbstractProductExportImportTestCase
      * @param string[] $skippedAttributes
      * @return void
      * @dataProvider exportImportDataProvider
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
     public function testImportExport(array $fixtures, array $skus, array $skippedAttributes = []): void
     {
-        $this->markTestSkipped('Uncomment after MAGETWO-38240 resolved');
+        $skippedAttributes = array_merge(self::$skippedAttributes, ['downloadable_links']);
+        parent::testImportExport($fixtures, $skus, $skippedAttributes);
     }
 
     /**
      * @inheritdoc
      */
     protected function assertEqualsSpecificAttributes(
-        \Magento\Catalog\Model\Product $expectedProduct,
-        \Magento\Catalog\Model\Product $actualProduct
+        Product $expectedProduct,
+        Product $actualProduct
     ): void {
-        $expectedProductLinks   = $expectedProduct->getExtensionAttributes()->getDownloadableProductLinks();
+        $expectedProductLinks = $expectedProduct->getExtensionAttributes()->getDownloadableProductLinks();
         $expectedProductSamples = $expectedProduct->getExtensionAttributes()->getDownloadableProductSamples();
 
-        $actualProductLinks   = $actualProduct->getExtensionAttributes()->getDownloadableProductLinks();
+        $actualProductLinks = $actualProduct->getExtensionAttributes()->getDownloadableProductLinks();
         $actualProductSamples = $actualProduct->getExtensionAttributes()->getDownloadableProductSamples();
 
         $this->assertEquals(count($expectedProductLinks), count($actualProductLinks));
         $this->assertEquals(count($expectedProductSamples), count($actualProductSamples));
+        $actualLinks = $this->getDataWithSortingById($actualProductLinks);
+        $expectedLinks = $this->getDataWithSortingById($actualProductLinks);
+        foreach ($actualLinks as $key => $actualLink) {
+            $this->assertEquals($expectedLinks[$key], $actualLink);
+        }
+        $actualSamples = $this->getDataWithSortingById($actualProductSamples);
+        $expectedSamples = $this->getDataWithSortingById($expectedProductSamples);
+        foreach ($actualSamples as $key => $actualSample) {
+            $this->assertEquals($expectedSamples[$key], $actualSample);
+        }
+    }
 
-        $expectedLinksArray = [];
-        foreach ($expectedProductLinks as $link) {
-            $expectedLinksArray[] = $link->getData();
-        }
-        foreach ($actualProductLinks as $actualLink) {
-            $this->assertContains($expectedLinksArray, $actualLink->getData());
+    /**
+     * Get data with sorting by id
+     *
+     * @param array $objects
+     *
+     * @return array
+     */
+    private function getDataWithSortingById(array $objects)
+    {
+        $result = [];
+        foreach ($objects as $object) {
+            $result[$object->getId()] = $object->getData();
         }
 
-        $expectedSamplesArray = [];
-        foreach ($expectedProductSamples as $sample) {
-            $expectedSamplesArray[] = $sample->getData();
-        }
-        foreach ($actualProductSamples as $actualSample) {
-            $this->assertContains($expectedSamplesArray, $actualSample->getData());
-        }
+        return $result;
     }
 }
