@@ -5,11 +5,22 @@
  */
 namespace Magento\Framework\App\Response\HeaderProvider;
 
+use Magento\Framework\App\Response\Http as HttpResponse;
+use Zend\Http\Header\HeaderInterface;
+
+/**
+ * Class AbstractHeaderTestCase
+ */
 abstract class AbstractHeaderTestCase extends \Magento\TestFramework\TestCase\AbstractController
 {
-    /** @var  \Magento\Framework\App\Response\Http */
+    /**
+     * @var HttpResponse
+     */
     private $interceptedResponse;
 
+    /**
+     * @inheritDoc
+     */
     public function setUp()
     {
         parent::setUp();
@@ -17,12 +28,11 @@ abstract class AbstractHeaderTestCase extends \Magento\TestFramework\TestCase\Ab
             [
                 'preferences' =>
                     [
-                        \Magento\Framework\App\Response\Http::class =>
-                            \Magento\Framework\App\Response\Http\Interceptor::class
+                        HttpResponse::class => 'Magento\Framework\App\Response\Http\Interceptor'
                     ]
             ]
         );
-        $this->interceptedResponse = $this->_objectManager->create(\Magento\Framework\App\Response\Http::class);
+        $this->interceptedResponse = $this->_objectManager->create(HttpResponse::class);
     }
 
     /**
@@ -33,16 +43,30 @@ abstract class AbstractHeaderTestCase extends \Magento\TestFramework\TestCase\Ab
      */
     protected function assertHeaderPresent($name, $value)
     {
+        $value = [$value];
         $this->interceptedResponse->sendResponse();
-
         $header = $this->interceptedResponse->getHeader($name);
-        $this->assertTrue(is_subclass_of($header, \Zend\Http\Header\HeaderInterface::class, false));
+
+        $headerContent = [];
+        if ($header instanceof \ArrayIterator) {
+            foreach ($header as $item) {
+                $headerContent[] = $item->getFieldValue();
+            }
+        } elseif ($header instanceof HeaderInterface) {
+            $headerContent[] = $header->getFieldValue();
+        }
+
         $this->assertSame(
             $value,
-            $header->getFieldValue()
+            $headerContent
         );
     }
 
+    /**
+     * Assert is no header.
+     *
+     * @param string $name
+     */
     protected function assertHeaderNotPresent($name)
     {
         $this->interceptedResponse->sendResponse();
