@@ -5,15 +5,19 @@
  */
 namespace Magento\Framework\View\Element;
 
-use Magento\Framework\View\Element\AbstractBlock;
+use Magento\Framework\Math\Random;
+use Magento\Framework\Session\SessionManagerInterface;
+use Magento\Framework\Session\SidResolverInterface;
+use Magento\TestFramework\Helper\Bootstrap;
 
 /**
  * @magentoAppIsolation enabled
+ * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  */
 class AbstractBlockTest extends \PHPUnit\Framework\TestCase
 {
     /**
-     * @var \Magento\Framework\View\Element\AbstractBlock
+     * @var AbstractBlock
      */
     protected $_block;
 
@@ -24,22 +28,29 @@ class AbstractBlockTest extends \PHPUnit\Framework\TestCase
 
     protected static $_mocks = [];
 
+    /**
+     * @var SessionManagerInterface
+     */
+    private $session;
+
     protected function setUp()
     {
-        \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->get(\Magento\Framework\App\State::class)
-            ->setAreaCode('frontend');
-        \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->get(
-            \Magento\Framework\View\DesignInterface::class
-        )->setDefaultDesignTheme();
+        /** @var \Magento\Framework\App\State $state */
+        $state = Bootstrap::getObjectManager()->get(\Magento\Framework\App\State::class);
+        $state->setAreaCode('frontend');
+        /** @var \Magento\Framework\View\DesignInterface $design */
+        $design = Bootstrap::getObjectManager()->get(\Magento\Framework\View\DesignInterface::class);
+        $design->setDefaultDesignTheme();
         $this->_block = $this->getMockForAbstractClass(
-            \Magento\Framework\View\Element\AbstractBlock::class,
+            AbstractBlock::class,
             [
-                \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->get(
+                Bootstrap::getObjectManager()->get(
                     \Magento\Framework\View\Element\Context::class
                 ),
                 ['module_name' => 'Magento_Theme']
             ]
         );
+        $this->session = Bootstrap::getObjectManager()->get(SessionManagerInterface::class);
     }
 
     /**
@@ -113,9 +124,9 @@ class AbstractBlockTest extends \PHPUnit\Framework\TestCase
         );
         $layout->createBlock(\Magento\Framework\View\Element\Template::class, $name);
         $block = $layout->getBlock($name);
-        $this->assertInstanceOf(\Magento\Framework\View\Element\AbstractBlock::class, $block);
+        $this->assertInstanceOf(AbstractBlock::class, $block);
         $block->setNameInLayout($name);
-        $this->assertInstanceOf(\Magento\Framework\View\Element\AbstractBlock::class, $layout->getBlock($name));
+        $this->assertInstanceOf(AbstractBlock::class, $layout->getBlock($name));
         $this->assertEquals($name, $block->getNameInLayout());
         $this->assertTrue($layout->hasElement($name));
         $newName = 'new_name';
@@ -549,7 +560,7 @@ class AbstractBlockTest extends \PHPUnit\Framework\TestCase
     public function testGetCacheKey()
     {
         $name = uniqid('block.');
-        $block = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->get(
+        $block = Bootstrap::getObjectManager()->get(
             \Magento\Framework\View\LayoutInterface::class
         )->createBlock(
             \Magento\Framework\View\Element\Text::class
@@ -562,6 +573,35 @@ class AbstractBlockTest extends \PHPUnit\Framework\TestCase
 
         $block->setCacheKey('key');
         $this->assertEquals(AbstractBlock::CACHE_KEY_PREFIX . 'key', $block->getCacheKey());
+    }
+
+    /**
+     * Check that SIDs inside blocks are not being replaced.
+     *
+     * @return void
+     * @magentoCache block_html enabled
+     */
+    public function testNoSid(): void
+    {
+        $blockId = 'block-with-sid' .Random::getRandomNumber(1, 9999);
+        $block = $this->_createBlockWithLayout(
+            $blockId,
+            $blockId,
+            \Magento\Framework\View\Element\Text::class
+        );
+        $outerId = 'block-outer' .Random::getRandomNumber(1, 9999);
+        $outer = $this->_createBlockWithLayout($outerId, $outerId);
+        $block->setText(
+            $text = 'Some text with ' .SidResolverInterface::SESSION_ID_QUERY_PARAM
+                .'=' .$this->session->getSessionId()
+        );
+        $block->setData('cache_lifetime', 3600);
+        //Caching the block's content
+        $outer->getBlockHtml($blockId);
+        //New ID generated, must not be replace in block's content.
+        $this->session->regenerateId();
+        $html = $outer->getBlockHtml($blockId);
+        $this->assertEquals($text, $html);
     }
 
     /**
@@ -610,7 +650,7 @@ class AbstractBlockTest extends \PHPUnit\Framework\TestCase
     protected function _createBlockWithLayout(
         $name = 'block',
         $alias = null,
-        $type = \Magento\Framework\View\Element\AbstractBlock::class
+        $type = AbstractBlock::class
     ) {
         $typePart = explode('\\', $type);
         $mockClass = array_pop($typePart) . 'Mock';
@@ -618,7 +658,7 @@ class AbstractBlockTest extends \PHPUnit\Framework\TestCase
             self::$_mocks[$mockClass] = $this->getMockForAbstractClass(
                 $type,
                 [
-                    \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->get(
+                    Bootstrap::getObjectManager()->get(
                         \Magento\Framework\View\Element\Context::class
                     ),
                     ['module_name' => 'Magento_Theme']
@@ -627,7 +667,7 @@ class AbstractBlockTest extends \PHPUnit\Framework\TestCase
             );
         }
         if ($this->_layout === null) {
-            $this->_layout = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()->get(
+            $this->_layout = Bootstrap::getObjectManager()->get(
                 \Magento\Framework\View\LayoutInterface::class
             );
         }
