@@ -12,13 +12,23 @@ require.config({
     }
 });
 
-define(['squire', 'ko', 'jquery', 'uiRegistry', 'jquery/validate'], function (Squire, ko, $, registry) {
+define(['squire', 'ko', 'jquery', 'jquery/validate'], function (Squire, ko, $) {
     'use strict';
 
     var injector = new Squire(),
         modalStub = {
             openModal: jasmine.createSpy(),
             closeModal: jasmine.createSpy()
+        },
+        country = {
+            /** Stub */
+            on: function () {},
+
+            /** Stub */
+            get: function () {},
+
+            /** Stub */
+            set: function () {}
         },
         mocks = {
             'Magento_Customer/js/model/customer': {
@@ -28,17 +38,7 @@ define(['squire', 'ko', 'jquery', 'uiRegistry', 'jquery/validate'], function (Sq
             'Magento_Checkout/js/model/address-converter': jasmine.createSpy(),
             'Magento_Checkout/js/model/quote': {
                 isVirtual: jasmine.createSpy(),
-                shippingMethod: ko.observable(),
-
-                /**
-                 * Stub
-                 */
-                shippingAddress: function () {
-
-                    return {
-                        'countryId': 'AD'
-                    };
-                }
+                shippingMethod: ko.observable()
             },
             'Magento_Checkout/js/action/create-shipping-address': jasmine.createSpy().and.returnValue(
                 jasmine.createSpyObj('newShippingAddress', ['getKey'])
@@ -62,7 +62,18 @@ define(['squire', 'ko', 'jquery', 'uiRegistry', 'jquery/validate'], function (Sq
                 'checkoutData',
                 ['setSelectedShippingAddress', 'setNewCustomerShippingAddress', 'setSelectedShippingRate']
             ),
-            'Magento_Ui/js/lib/registry/registry': registry,
+            'Magento_Ui/js/lib/registry/registry': {
+                async: jasmine.createSpy().and.returnValue(function () {}),
+                create: jasmine.createSpy(),
+                set: jasmine.createSpy(),
+                get: jasmine.createSpy().and.callFake(function (query) {
+                    if (query === 'test.shippingAddress.shipping-address-fieldset.country_id') {
+                        return country;
+                    } else if (query === 'checkout.errors') {
+                        return {};
+                    }
+                })
+            },
             'Magento_Checkout/js/model/shipping-rate-service': jasmine.createSpy()
         },
         obj;
@@ -73,7 +84,6 @@ define(['squire', 'ko', 'jquery', 'uiRegistry', 'jquery/validate'], function (Sq
             obj = new Constr({
                 provider: 'provName',
                 name: '',
-                parentName: 'test',
                 index: '',
                 popUpForm: {
                     options: {
@@ -174,16 +184,6 @@ define(['squire', 'ko', 'jquery', 'uiRegistry', 'jquery/validate'], function (Sq
 
         describe('"validateShippingInformation" method', function () {
             it('Check method call on negative cases.', function () {
-                /* jscs:disable */
-                var country = {
-                    on: function () {},
-                    get: function () {},
-                    set: function () {}
-                };
-                /* jscs:enable */
-
-                registry.set('test.shippingAddress.shipping-address-fieldset.country_id', country);
-                registry.set('checkout.errors', {});
                 obj.source = {
                     get: jasmine.createSpy().and.returnValue(true),
                     set: jasmine.createSpy(),
@@ -199,20 +199,10 @@ define(['squire', 'ko', 'jquery', 'uiRegistry', 'jquery/validate'], function (Sq
                 expect(obj.validateShippingInformation()).toBeFalsy();
             });
             it('Check method call on positive case.', function () {
-                /* jscs:disable */
-                var country = {
-                    on: function () {},
-                    get: function () {},
-                    set: function () {}
-                };
-                /* jscs:enable */
-
                 $('body').append('<form data-role="email-with-possible-login">' +
                     '<input type="text" name="username" />' +
                     '</form>');
 
-                registry.set('test.shippingAddress.shipping-address-fieldset.country_id', country);
-                registry.set('checkout.errors', {});
                 obj.source = {
                     get: jasmine.createSpy().and.returnValue(true),
                     set: jasmine.createSpy(),
