@@ -3,30 +3,50 @@
  * See COPYING.txt for license details.
  */
 
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+
 define([
-    'underscore',
-    'uiRegistry',
-    'Magento_Ui/js/form/form'
-], function (_, registry, Constr) {
+    'squire'
+], function (Squire) {
     'use strict';
 
     describe('Magento_Ui/js/form/form', function () {
+        var injector = new Squire(),
+            mocks = {
+                'Magento_Ui/js/lib/registry/registry': {
+                    /** Method stub. */
+                    get: function () {
+                        return {
+                            get: jasmine.createSpy(),
+                            set: jasmine.createSpy()
+                        };
+                    },
+                    options: jasmine.createSpy(),
+                    create: jasmine.createSpy(),
+                    set: jasmine.createSpy(),
+                    async: jasmine.createSpy()
+                }
+            },
+            obj,
+            dataScope = 'dataScope';
 
-        var obj = new Constr({
-            provider: 'provName',
-            name: '',
-            index: ''
-        });
+        beforeEach(function (done) {
+            injector.mock(mocks);
+            injector.require([
+                'Magento_Ui/js/form/form'
+            ], function (Constr) {
+                obj = new Constr({
+                    provider: 'provName',
+                    name: '',
+                    index: '',
+                    dataScope: dataScope
+                });
 
-        registry.set('provName', {
-            /** Stub */
-            on: function () {},
-
-            /** Stub */
-            get: function () {},
-
-            /** Stub */
-            set: function () {}
+                done();
+            });
         });
 
         describe('"initAdapter" method', function () {
