@@ -10,7 +10,8 @@ namespace Magento\Framework\Session;
 use Magento\Framework\App\State;
 
 /**
- * Class SidResolver
+ * Resolves SID by processing request parameters.
+ *
  * @deprecated 2.3.3 SIDs in URLs are no longer used
  */
 class SidResolver implements SidResolverInterface
@@ -27,11 +28,13 @@ class SidResolver implements SidResolverInterface
 
     /**
      * @var \Magento\Framework\UrlInterface
+     * @deprecated Not used anymore.
      */
     protected $urlBuilder;
 
     /**
      * @var \Magento\Framework\App\RequestInterface
+     * @deprecated Not used anymore.
      */
     protected $request;
 
@@ -61,17 +64,13 @@ class SidResolver implements SidResolverInterface
     protected $_scopeType;
 
     /**
-     * @var State
-     */
-    private $appState;
-
-    /**
      * @param \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
      * @param \Magento\Framework\UrlInterface $urlBuilder
      * @param \Magento\Framework\App\RequestInterface $request
      * @param string $scopeType
      * @param array $sidNameMap
      * @param State|null $appState
+     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
     public function __construct(
         \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
@@ -86,34 +85,19 @@ class SidResolver implements SidResolverInterface
         $this->request = $request;
         $this->sidNameMap = $sidNameMap;
         $this->_scopeType = $scopeType;
-        $this->appState = $appState ?: \Magento\Framework\App\ObjectManager::getInstance()->get(State::class);
     }
 
     /**
      * Get Sid
      *
      * @param SessionManagerInterface $session
-     *
      * @return string|null
      * @throws \Magento\Framework\Exception\LocalizedException
+     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
     public function getSid(SessionManagerInterface $session)
     {
-        if ($this->appState->getAreaCode() !== \Magento\Framework\App\Area::AREA_FRONTEND) {
-            return null;
-        }
-
-        $sidKey = null;
-
-        $useSidOnFrontend = $this->getUseSessionInUrl();
-        if ($useSidOnFrontend && $this->request->getQuery(
-            $this->getSessionIdQueryParam($session),
-            false
-        ) && $this->urlBuilder->isOwnOriginUrl()
-        ) {
-            $sidKey = $this->request->getQuery($this->getSessionIdQueryParam($session));
-        }
-        return $sidKey;
+        return null;
     }
 
     /**
