@@ -78,7 +78,7 @@ define(function () {
             }
         }
 
-        //Check if document already complete, and if so, just trigger page load
+        //Check if document is no longer loading, and if so, just trigger page load
         //listeners. Latest webkit browsers also use "interactive", and
         //will fire the onDOMContentLoaded before "interactive" but not after
         //entering "interactive" or "complete". More details:
@@ -89,8 +89,9 @@ define(function () {
         //so removing the || document.readyState === "interactive" test.
         //There is still a window.onload binding that should get fired if
         //DOMContentLoaded is missed.
-        if (document.readyState === "complete") {
-            pageLoaded();
+        if (document.readyState !== "loading") {
+            // Handle it asynchronously to allow scripts the opportunity to delay ready
+            setTimeout(pageLoaded);
         }
     }
 
